Issue #6: refactored gridview and updated navbar
