+++ lib/screens/navbar_android.dart
@@ -4,6 +4,7 @@ import 'package:google_nav_bar/google_nav_bar.dart';
 import 'package:flutter/material.dart';
 import 'login_page.dart';
 import 'search_bar.dart';
+import 'gridview.dart';
 
 class MyHomePageApp extends StatefulWidget{
   const MyHomePageApp({Key? key}): super(key: key);
@@ -12,162 +13,7 @@ class MyHomePageApp extends StatefulWidget{
   State<MyHomePageApp> createState() => _MyHomePageAppState();
 }
 
-class GridB extends StatefulWidget {
-  const GridB({super.key});
 
-  @override
-  State<GridB> createState() => _GridBState();
-}
-
-class _GridBState extends State<GridB> {
-  final List<Map<String, dynamic>> gridMap =[
-    {
-      "name": "Cheeseburger",
-      "rating" : "4.5",
-      "images" : "https://images.unsplash.com/photo-1586190848861-99aa4a171e90?q=80&w=2080&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
-      
-    },
-    {
-      "name": "Cheeseburger",
-      "rating" : "4.5",
-      "images" : "https://images.unsplash.com/photo-1586190848861-99aa4a171e90?q=80&w=2080&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
-      
-    },{
-      "name": "Cheeseburger",
-      "rating" : "4.5",
-      "images" : "https://images.unsplash.com/photo-1586190848861-99aa4a171e90?q=80&w=2080&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
-      
-    },
-    {
-      "name": "Cheeseburger",
-      "rating" : "4.5",
-      "images" : "https://images.unsplash.com/photo-1586190848861-99aa4a171e90?q=80&w=2080&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
-      
-    },
-    {
-      "name": "Cheeseburger",
-      "rating" : "4.5",
-      "images" : "https://images.unsplash.com/photo-1586190848861-99aa4a171e90?q=80&w=2080&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
-      
-    },
-    {
-      "name": "Cheeseburger",
-      "rating" : "4.5",
-      "images" : "https://images.unsplash.com/photo-1586190848861-99aa4a171e90?q=80&w=2080&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
-      
-    },
-    {
-      "name": "Cheeseburger",
-      "rating" : "4.5",
-      "images" : "https://images.unsplash.com/photo-1586190848861-99aa4a171e90?q=80&w=2080&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
-      
-    },
-    {
-      "name": "Cheeseburger",
-      "rating" : "4.5",
-      "images" : "https://images.unsplash.com/photo-1586190848861-99aa4a171e90?q=80&w=2080&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
-      
-    },
-    {
-      "name": "Cheeseburger",
-      "rating" : "4.5",
-      "images" : "https://images.unsplash.com/photo-1586190848861-99aa4a171e90?q=80&w=2080&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
-      
-    },
-
-  ];
-
-Widget build(BuildContext context) {
-    return GridView.builder(
-      physics: const NeverScrollableScrollPhysics(),
-      shrinkWrap: true,
-      gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
-        crossAxisCount: 2,
-        crossAxisSpacing: 12.0,
-        mainAxisSpacing: 12.0,
-        mainAxisExtent: 310,
-      ),
-      itemCount: gridMap.length,
-      itemBuilder: (_, index) {
-        return Container(
-          decoration: BoxDecoration(
-            borderRadius: BorderRadius.circular(
-              16.0,
-            ),
-            color: Colors.black,
-          ),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              ClipRRect(
-                borderRadius: const BorderRadius.only(
-                  topLeft: Radius.circular(16.0),
-                  topRight: Radius.circular(16.0),
-                ),
-                child: Image.network(
-                  "${gridMap.elementAt(index)['images']}",
-                  height: 170,
-                  width: double.infinity,
-                  fit: BoxFit.cover,
-                ),
-              ),
-              Padding(
-                padding: const EdgeInsets.all(8.0),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      "${gridMap.elementAt(index)['name']}",
-                      style: Theme.of(context).textTheme.subtitle1!.merge(
-                            const TextStyle(
-                              fontWeight: FontWeight.w700,
-                              color: Colors.white,
-                            ),
-                          ),
-                    ),
-                    const SizedBox(
-                      height: 8.0,
-                    ),
-                    Text(
-                      "${gridMap.elementAt(index)['rating']}",
-                      style: Theme.of(context).textTheme.subtitle2!.merge(
-                            TextStyle(
-                              fontWeight: FontWeight.w700,
-                              color: Colors.grey.shade500,
-                            ),
-                          ),
-                    ),
-                    const SizedBox(
-                      height: 8.0,
-                    ),
-                    Row(
-                      children: [
-                        IconButton(
-                          onPressed: () {},
-                          color: Colors.pink,
-                          icon: const Icon(
-                            CupertinoIcons.heart,
-                          ),
-                        ),
-                        IconButton(
-                          onPressed: () {},
-                          color: Colors.white,
-                          icon: const Icon(
-                            CupertinoIcons.search,
-                          ),
-                        ),
-                      ],
-                    ),
-                  ],
-                ),
-              ),
-            ],
-          ),
-        );
-      },
-    );
-  }
-}
 
 class _MyHomePageAppState extends State<MyHomePageApp>{
   @override
@@ -178,8 +24,6 @@ class _MyHomePageAppState extends State<MyHomePageApp>{
          child: SingleChildScrollView(
            child: Column(
              children: [
-               
-                 
                    Container(
                     constraints: BoxConstraints(maxHeight: 90.0),
                      child: Padding(
@@ -188,16 +32,10 @@ class _MyHomePageAppState extends State<MyHomePageApp>{
                      ),
                    ),
                    //SizedBox(height: 1000,),
-                 
-               
-               
-                 
                    Padding(
                      padding: EdgeInsets.all(8.0),
                      child: GridB(),
                    ),
-                 
-               
              ],
            ),
          ),
